Refactored draftOfPhpJsonUtilities.php
Merged the JsonString and JsonSerializedObject classes, it makes
more sense to have a single class to represent a JsonString for now,
and makes it easier to control/predict how the encoding of various
types is handled.

Put another way, it could become very difficult to implement a
standard encoding/decoding if their are different types of
JsonString. In that context, a single decoder would have to account
for the possibility that 2 JsonString implementations may encode
values differently, and would potentially require the implementation
of multiple decoders to handle various types of JsonString.

The actual JsonString and JsonDecoder classes should be implemented
as final classes, and provide a standard way to encode/decode values
as json.

This is a working draft.