Working on `draftOfPhpJsonUtilities.php`
Fixed value of `Mocker::INTEGER`, was `integer`, is now `int`.

Refactored scope of `Mocker->reflectionClass()` to `private`.

Default value for strings returned by
`Mocker->generateMockClassMethodArguments()` is no longer a
randomly generated string of characters. Instead a `SafeText`
instance is used to generate the default string.

The `Mocker->generateMockClassMethodArguments()` method
now throws an error if any arguments could not be mocked.

`JsonStrings` no longer allow the encoding of a `Closure`.

`JsonStrings` no longer allow the encoding of a `Directory`.

Removed `hack` code that functioned as a temporary workaround to
allow encoding/decoding of a `JsonSting`. `JsonSting`s are no
longer allowed to be encoded as `JsonString`s.

Defined new test classes.