Resolved issues that caused the following test to fail:
```
JsonDecoderTestTrait->test_decode_returns_the_original_data()

```

The issue:

If an object that had uninitialized properties was encoded as Json,
and an attempt was made to decode it via `JsonDecoder->decode()`,
the JsonDecoder would attempt to assign `null` to the property.
If the property was not nullable, this resulted in an error.